create api, make conda optional
